Trade Ledger V1 and UI update [structures and APIs]
Two savegame version bumps - beta version issued to the testing team had to be expanded. The mid-step version is therefore preserved to ensure compatibility, in case bugs from the testing version make it onto the master.

+++ src/empire/editor.c
@@ -79,8 +79,8 @@ void empire_editor_set_tool(empire_tool tool)
 
 void empire_editor_change_tool(int amount)
 {
-    if ((int)data.current_tool + amount < EMPIRE_TOOL_MIN) {
-        data.current_tool = EMPIRE_TOOL_MAX - ((int)data.current_tool - amount) + 1;
+    if ((int) data.current_tool + amount < EMPIRE_TOOL_MIN) {
+        data.current_tool = EMPIRE_TOOL_MAX - ((int) data.current_tool - amount) + 1;
     } else {
         data.current_tool += amount;
     }
@@ -155,7 +155,7 @@ static void shift_edge_indices(const empire_object *const_obj)
     if (const_obj->order_index < data.foreach_param1) {
         return;
     }
-    if (const_obj->parent_object_id != empire_object_get_border()->id) {
+    if ((unsigned int) const_obj->parent_object_id != empire_object_get_border()->id) {
         return;
     }
     empire_object *obj = empire_object_get(const_obj->id);
@@ -321,7 +321,8 @@ static int place_object(int mouse_x, int mouse_y)
     return 1;
 }
 
-static int create_trade_route_default(full_empire_object *full) {
+static int create_trade_route_default(full_empire_object *full)
+{
     full_empire_object *route_obj = empire_object_get_new();
     if (!route_obj) {
         log_error("Error creating new object - out of memory", 0, 0);
@@ -339,7 +340,7 @@ static int place_city(full_empire_object *city_obj)
 {
     city_obj->in_use = 1;
     city_obj->obj.type = EMPIRE_OBJECT_CITY;
-    
+
     switch (data.current_tool) {
         case EMPIRE_TOOL_OUR_CITY:
             if (empire_object_get_our_city()) {
@@ -391,7 +392,7 @@ static int place_city(full_empire_object *city_obj)
             city_obj->empire_city_icon = EMPIRE_CITY_ICON_DISTANT_CITY;
             break;
         default:
-            return 0; 
+            return 0;
     }
 
     city_obj->obj.image_id = empire_city_get_icon_image_id(city_obj->empire_city_icon);
@@ -407,7 +408,7 @@ static int place_city(full_empire_object *city_obj)
 static int place_border(full_empire_object *edge)
 {
     edge->in_use = 1;
-    unsigned int parent_id = 0; 
+    unsigned int parent_id = 0;
     const empire_object *current_border = empire_object_get_border();
     if (!current_border) {
         // create border
@@ -446,7 +447,7 @@ static int place_battle(full_empire_object *battle_obj)
     const image *img = image_get(battle_obj->obj.image_id);
     battle_obj->obj.width = img->width;
     battle_obj->obj.height = img->height;
-    
+
     return 1;
 }
 
@@ -471,7 +472,7 @@ static int place_distant_battle(full_empire_object *distant_battle)
     const image *img = image_get(distant_battle->obj.image_id);
     distant_battle->obj.width = img->width;
     distant_battle->obj.height = img->height;
-    
+
     return 1;
 }
 
@@ -511,7 +512,7 @@ static void deletion_confirmed(int confirmed, int checked)
 }
 
 static int delete_object_at(int mouse_x, int mouse_y)
-{    
+{
     int empire_x = editor_empire_mouse_to_empire_x(mouse_x);
     int empire_y = editor_empire_mouse_to_empire_y(mouse_y);
     data.deletion_id = empire_object_get_at(empire_x, empire_y);
@@ -574,7 +575,7 @@ int empire_editor_delete_object(unsigned int obj_id)
 }
 
 static int pick_empire_tool(int mouse_x, int mouse_y)
-{    
+{
     int empire_x = editor_empire_mouse_to_empire_x(mouse_x);
     int empire_y = editor_empire_mouse_to_empire_y(mouse_y);
     int picked_id = empire_object_get_at(empire_x, empire_y);
@@ -584,7 +585,7 @@ static int pick_empire_tool(int mouse_x, int mouse_y)
     full_empire_object *full = empire_object_get_full(picked_id);
     empire_tool object_tool = empire_editor_get_tool_for_object(full);
     if (!object_tool && full->city_type != EMPIRE_CITY_OURS) {
-        return 0;   
+        return 0;
     }
     data.current_tool = object_tool;
     window_request_refresh();
